search/gvg_select/requirements.txt
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-bootstrap-components==1.7.1
gunicorn==22.0.0
folium==0.16.0
pandas==2.2.3
numpy==2.1.1
psycopg[binary]==3.2.3
psycopg2-binary==2.9.9; python_version < "3.13"
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.2
Unidecode==1.3.8